Extract Report module with thin views for season/tournament end
Move season-end and tournament-end logic from fat view classes into
dedicated service classes under a new app/Modules/Report/ module:
- SeasonSummaryService (from ShowSeasonEnd)
- CompetitionSummaryService (from ShowTournamentEnd)
- AwardService (shared award queries)

Views are now thin orchestrators that delegate to the services.
